refactor: keep only what production uses in the line grid module
Nothing in src/ decodes a line grid descriptor — the shader does that itself,
in GLSL — so decodeLineGrid and findBands existed solely for the tests, and the
format constants were exported solely so the tests could assert the layout.
They now live in tests/line-grid.js, and the constants go back to being private
to the module that encodes them.

findBands was the last caller of decodeLineGrid, and was itself built for the
JavaScript upscaler that 51adf18 deleted; it outlived what it was for.

Putting the decoder in the tests also makes them independent: they reimplement
the inverse rather than reusing the encoder's own helper, so a wrong bit layout
fails instead of cancelling itself out.
