+++ tests/unit/test-6847.js
@@ -1,6 +1,6 @@
 import { describe, it, expect } from "vitest";
 import { Video6847 } from "../../src/6847.js";
-import { decodeLineGrid } from "../../src/video-filters/pixel-grid.js";
+import { decodeLineGrid } from "../line-grid.js";
 import { shortestRun } from "../pixel-runs.js";
 
 // The Video6847 constructor needs a `video` object with a few methods it
