+++ tests/unit/test-video.js
@@ -1,7 +1,8 @@
 import { describe, it, expect, beforeEach, vi, afterEach } from "vitest";
 import { Video, HDISPENABLE, VDISPENABLE, USERDISPENABLE, EVERYTHINGENABLED } from "../../src/video.js";
 import * as utils from "../../src/utils.js";
-import { decodeLineGrid, texelsPerPixel } from "../../src/video-filters/pixel-grid.js";
+import { texelsPerPixel } from "../../src/video-filters/pixel-grid.js";
+import { decodeLineGrid } from "../line-grid.js";
 import { shortestRun } from "../pixel-runs.js";
 
 // Setup with focus on testing behavior rather than implementation details
